add app name in status
merge same app status

not set raw data when hit suspend

support tlp in chart

clear rendered code file

support set in rendered
